Move logger to shared
