--- packages/observability-node/README.md
@@ -1,3 +1,0 @@
-# @replayio/observability-node
-
-Node utilities for Replay.io's own observability.
