--- packages/observability-node/tsconfig.json
@@ -1,5 +1,0 @@
-{
-  "extends": "@replay-cli/tsconfig/base.json",
-  "compilerOptions": {},
-  "include": ["src/**/*.ts"]
-}
