+++ packages/shared/package.json
@@ -3,11 +3,20 @@
   "version": "0.0.0",
   "private": true,
   "exports": {
+    "./logger": "./dist/logger/index.js",
     "./*": "./dist/*.js"
   },
   "@replay-cli/pkg-build": {
     "entrypoints": [
-      "./src/*.ts"
+      "./src/*.ts",
+      "./src/logger/index.ts"
     ]
+  },
+  "dependencies": {
+    "debug": "^4.3.4",
+    "fs-extra": "^11.2.0",
+    "undici": "^5.28.4",
+    "winston": "^3.13.0",
+    "winston-loki": "^6.1.2"
   }
 }
