+++ packages/shared/src/logger/config.ts
@@ -1,3 +1,1 @@
-// TODO [PRO-629] Move this into the "shared" package.
-
 export const replayApiServer = process.env.REPLAY_API_SERVER || "https://api.replay.io";
