+++ packages/shared/src/logger/graphql/GraphQLError.ts
@@ -1,5 +1,3 @@
-// TODO [PRO-629] Move this into the "shared" package.
-
 export class GraphQLError extends Error {
   errors: unknown[];
