+++ packages/shared/src/logger/graphql/cachePath.ts
@@ -1,3 +1,3 @@
-import { getReplayPath } from "../getReplayPath";
+import { getReplayPath } from "../../getReplayPath";
 
 export const cachePath = getReplayPath("observability-profile", "graphql.json");
