+++ packages/shared/src/logger/graphql/updateCachedAuthIds.ts
@@ -1,5 +1,3 @@
-// TODO [PRO-629] Move this into the "shared" package.
-
 import { readFromCache, writeToCache } from "../cache";
 import { cachePath } from "./cachePath";
 import { AuthIds } from "./fetchAuthIdsFromGraphQL";
