--- packages/test-utils/src/logger/getUserAgent.ts
@@ -1,7 +1,0 @@
-// TODO [PRO-629] Move this into the "shared" package.
-
-import { name, version } from "../../package.json";
-
-export function getUserAgent() {
-  return `${name}/${version}`;
-}
