+++ packages/test-utils/tsconfig.json
@@ -9,6 +9,9 @@
   "references": [
     {
       "path": "../replay"
+    },
+    {
+      "path": "../shared"
     }
   ]
 }
