+++ tsconfig.json
@@ -3,7 +3,6 @@
   "references": [
     { "path": "./packages/cypress" },
     { "path": "./packages/jest" },
-    { "path": "./packages/observability-node" },
     { "path": "./packages/playwright" },
     { "path": "./packages/puppeteer" },
     { "path": "./packages/replay" },
